Move all of signaling code to rtcpeerconnection.go file
